// web/modules/component-examples/components/ComponentExamplesForm/ComponentExamplesForm.enums.ts
export enum EComponentExampleRole {
  ADMIN = "admin",
  EDITOR = "editor",
  VIEWER = "viewer",
}

export enum EComponentExamplePlan {
  FREE = "free",
  PRO = "pro",
  ENTERPRISE = "enterprise",
}

export enum EComponentExampleTag {
  DESIGN = "design",
  ENGINEERING = "engineering",
  PRODUCT = "product",
  MARKETING = "marketing",
}